Fixed dependency system of Echo, changed version of core,client and server to 0.2.0, added LICENSE for Echo

+++ core/src/EchoWriter.cpp
@@ -7,7 +7,7 @@ echo::EchoWriter *echo::EchoWriter::instance = nullptr;
     if (i > deadLock)
       return -1;
     int wroteBytes = ::write(sock, buffer, bufferLen);
-    if (wroteBytes < bufferLen) {
+    if (wroteBytes < (int)bufferLen) {
       /*d - delta*/
       int d = bufferLen - wroteBytes;
       write(sock, buffer + wroteBytes, d, i++);
